% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{NMAPower-package}
\alias{NMAPower}
\alias{NMAPower-package}
\title{NMAPower package}
\description{
Simulation-based power analysis for Bayesian network meta-analysis.
}
\author{
\strong{Maintainer}: Yicheng Shen \email{yshen128@jh.edu}

Authors:
\itemize{
  \item Yicheng Shen \email{yshen128@jh.edu}
}

}
\keyword{internal}
