% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_power_sim.R
\name{print.nma_power_sim}
\alias{print.nma_power_sim}
\title{Print method for nma_power_sim objects}
\usage{
\method{print}{nma_power_sim}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{nma_power_sim()}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print method for nma_power_sim objects
}
