% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_power_sim_3node.R
\name{print.nma_power_sim_3node}
\alias{print.nma_power_sim_3node}
\title{Print method for nma_power_sim_3node objects}
\usage{
\method{print}{nma_power_sim_3node}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{nma_power_sim_3node()}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print method for nma_power_sim_3node objects
}
