% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_power_sim.R, R/nma_power_sim_3node.R
\name{print.nma_power_sim}
\alias{print.nma_power_sim}
\title{Print method for nma_power_sim objects}
\usage{
\method{print}{nma_power_sim}(x, ...)

\method{print}{nma_power_sim}(x, ...)
}
\description{
Print method for nma_power_sim objects

Print method for nma_power_sim objects
}
